backend/requirements.txt
# can add manually
# or
# add using"pip freeze  > requirements.txt" -- this one freeze packages exact with the versions

fastapi
uvicorn[standard]
pymongo
python-dotenv
